+++ QCM_main/QCM_main.py
@@ -79,11 +79,30 @@ class VNATracker:
         self.avg = None     # average of scans (int)
         self.speed = None   # vna speed set up (int 1 to 10)
         self.instrmode = 0  # instrument mode (0: reflection)
+        self.cal = self.get_cal_filenames()
         
         self.setflg = {} # if vna needs to reset (set with reset selections)
         self.setflg.update(self.__dict__) # get all attributes in a dict
         self.setflg.pop('setflg', None) # remove setflg itself
-    
+           
+    def get_cal_filenames(self):
+        '''
+        find calc file for ADC1 and ADC2 separately
+        The fill should be stored in settings_init['vna_cal_file_path'] for each channel 
+        ''' 
+        cal = {'ADC1': '', 'ADC2': ''}
+        if (UIModules.system_check() == 'win32') and (struct.calcsize('P') * 8 == 32): # windows (if is win32, struct will already be imported above)
+            for key in cal.keys():
+                files = os.listdir(settings_init['vna_cal_file_path']) # list all file in the given folder
+                for file in files:
+                    if (key + '.myVNA.cal').lower() in file.lower():
+                        cal[key] = os.path.abspath(os.path.join(settings_init['vna_cal_file_path'], file)) # use absolute path
+                        break
+
+        return cal
+
+
+
     def set_check(self, **kwargs):
         for key, val in kwargs.items():
             if isinstance(val, np.ndarray): # self.f 
@@ -689,7 +708,7 @@ class QCMApp(QMainWindow):
 
         # self.ui.checkBox_settings_temp_sensor.stateChanged.connect(self.update_tempsensor)
         self.ui.checkBox_settings_temp_sensor.stateChanged.connect(self.on_clicked_set_temp_sensor)
-        self.ui.comboBox_thrmcpltype.currentIndexChanged.connect(self.update_tempdevice)
+        # self.ui.comboBox_thrmcpltype.currentIndexChanged.connect(self.update_tempdevice) # ??
         self.ui.comboBox_thrmcpltype.currentIndexChanged.connect(self.update_thrmcpltype)
 
         # set signals to update plots settings_settings
@@ -1392,9 +1411,33 @@ class QCMApp(QMainWindow):
 
     def on_triggered_actionOpen_MyVNA(self):
         '''
-        open myVNA
+        open myVNA.exe
         '''
-        subprocess.Popen(settings_init['vna_path'])
+        if UIModules.system_check() != 'win32': # not windows
+            return
+
+        myvna_path = self.settings.get('vna_path', '')
+        if myvna_path and os.path.exists(myvna_path): # user defined myVNA.exe path exists and correct
+            pass
+        else: # use default path list
+            for myvna_path in settings_init['vna_path']:
+                if os.path.exists(myvna_path):
+                    break
+                else:
+                    myvna_path = ''
+        
+        if myvna_path:
+            subprocess.call(myvna_path) # open myVNA
+        else:
+            process = self.process_messagebox(
+                text='Failed to open myVNA.exe',
+                message=['Cannot find myVNA.exe in: \n{}\nPlease add the path for "vna_path" in "settings_default.json"!'.format('\n'.join(settings_init['vna_path'])),
+                'The format of the path should like this:',
+                r'"C:\\Program Files (x86)\\G8KBB\\myVNA\\myVNA.exe"'
+                ],
+                opts=False, 
+                forcepop=True,
+            )
 
 
     def on_triggered_actionImport_QCM_D(self):
@@ -1593,6 +1636,7 @@ class QCMApp(QMainWindow):
                 with open(fileName, 'w') as f:
                     settings = self.settings.copy()
                     settings.pop('dateTimeEdit_reftime', None)
+                    settings.pop('dateTimeEdit_settings_data_t0shifted', None)
                     line = json.dumps(settings, indent=4) + "\n"
                     f.write(line)
                 print('Settings were exported as json file.')
@@ -1650,7 +1694,7 @@ class QCMApp(QMainWindow):
         if fileName:
             self.data_saver.data_exporter(fileName) # do the export
 
-    def process_messagebox(self, message=[], forcepop=False):
+    def process_messagebox(self, text='Your selection was paused!', message=[], opts=True, forcepop=False):
         '''
         check is the experiment is ongoing (self.timer.isActive()) and if data is saved (self.data_saver.saveflg)
         and pop up a messageBox to ask if process
@@ -1669,12 +1713,15 @@ class QCMApp(QMainWindow):
                 message.append('Test is Running!')
                 buttons = QMessageBox.Ok
             else:
-                message.append('Do you want to process?')
-                buttons = QMessageBox.Yes | QMessageBox.Cancel
+                if not opts:
+                    buttons = QMessageBox.Ok
+                else:
+                    message.append('Do you want to process?')
+                    buttons = QMessageBox.Yes | QMessageBox.Cancel
 
             msg = QMessageBox()
             msg.setIcon(QMessageBox.Information)
-            msg.setText('Your selection was paused!')
+            msg.setText(text)
             msg.setInformativeText('\n'.join(message))
             msg.setWindowTitle(_version.__projectname__ + ' Message')
             msg.setStandardButtons(buttons)
@@ -2055,10 +2102,15 @@ class QCMApp(QMainWindow):
         self.ui.mpl_spectra_fit.update_data({'ln': 'lB', 'x': f, 'y': B})
 
         # constrain xlim
-        self.ui.mpl_spectra_fit.ax[0].set_xlim(f[0], f[-1])
-        self.ui.mpl_spectra_fit.ax[1].set_xlim(f[0], f[-1])
-        self.ui.mpl_spectra_fit.ax[0].set_ylim(min(G)-0.05*(max(G)-min(G)), max(G)+0.05*(max(G)-min(G)))
-        self.ui.mpl_spectra_fit.ax[1].set_ylim(min(B)-0.05*(max(B)-min(B)), max(B)+0.05*(max(B)-min(B)))
+        if (f is not None) and (f[0] != f[-1]): # f is available
+            self.ui.mpl_spectra_fit.ax[0].set_xlim(f[0], f[-1])
+            self.ui.mpl_spectra_fit.ax[1].set_xlim(f[0], f[-1])
+            self.ui.mpl_spectra_fit.ax[0].set_ylim(min(G)-0.05*(max(G)-min(G)), max(G)+0.05*(max(G)-min(G)))
+            self.ui.mpl_spectra_fit.ax[1].set_ylim(min(B)-0.05*(max(B)-min(B)), max(B)+0.05*(max(B)-min(B)))
+        elif f is None or (not f.any()): # vna error or f is all 0s
+            self.ui.mpl_spectra_fit.ax[0].autoscale()
+            self.ui.mpl_spectra_fit.ax[1].autoscale()
+
 
         ## connect axes event
         self.mpl_connect_cid(self.ui.mpl_spectra_fit, self.on_fit_lims_change)
@@ -3106,6 +3158,9 @@ class QCMApp(QMainWindow):
             return
         # get f1
         f1 = self.data_saver.get_fg_ref(chn_name, harms=[1])['f0'][0]
+        if np.isnan(f1): # no 1st harmonic data collected
+            # use base frequency in settings
+            f1 = float(self.settings['comboBox_base_frequency']) * 1e6 # in Hz
         # set f1 to qcm module
         self.qcm.f1 = f1
 
@@ -3457,8 +3512,8 @@ class QCMApp(QMainWindow):
                 thrmcpltype = self.settings['comboBox_thrmcpltype'] # get thermocouple type
                 tempdevice = TempDevices.device_info(self.settings['comboBox_tempdevice']) #get temp device info
 
-                # check senor availability
-                package_str = settings_init['tempmodules_path'][2:].replace('/', '.') + tempmodule_name
+                # # check senor availability
+                # package_str = settings_init['tempmodules_path'][2:].replace('/', '.') + tempmodule_name
                 # import package
                 temp_sensor = getattr(TempModules, tempmodule_name)
 
@@ -4526,6 +4581,23 @@ class QCMApp(QMainWindow):
                     # get data
                     f[chn_name][harm], G[chn_name][harm], B[chn_name][harm] = self.get_vna_data_no_with(harm=harm, chn_name=chn_name)
                     
+                    print('check:')
+                    print(f[chn_name][harm] is None)
+                    print(f[chn_name][harm][0] == f[chn_name][harm][-1])
+                    if (f[chn_name][harm] is None) or (f[chn_name][harm][0] == f[chn_name][harm][-1]): # vna error
+                        print('Can''t find analyzer!')
+                        # stop test
+                        self.idle = True
+                        self.ui.pushButton_runstop.setChecked(False)
+                        # alert
+                        process = self.process_messagebox(
+                            text='Failed to connect with analyzer!',
+                            message=['Please checked the connection with analyzer or if it''s power is on.'],
+                            opts=False, 
+                            forcepop=True,
+                        )
+                        return
+
                     # put f, G, B to peak_tracker for later fitting and/or tracking
                     self.peak_tracker.update_input(chn_name, harm, f[chn_name][harm], G[chn_name][harm], B[chn_name][harm], self.settings['harmdata'], self.settings['freq_span'])
 
@@ -4648,10 +4720,6 @@ class QCMApp(QMainWindow):
 
 
         self.idle = True
-
-
-        self.writing = True
-        # save scans to file
 
         self.writing = False