+++ CHANGELOG.md
@@ -1,9 +1,25 @@
 # Changelog
 
-## [0.15.0] - 2019-01-15
+## [0.15.1] - 2019-01-18
+
 ### Added
+
+- Add set frequency range to myVNA for display data in it, in case it is needed to check data out of Python program.
+- Add base frequency check for property claculation in UI. Now, the base frequency set in UI will be used as f1 when first harmonic data is not collected.
+- Add load calibration file when switch channel
+
+### Fixed
+
+- Fix program crashes when collecting data with VNA disconnected on Windows  
+  
+## [0.15.0] - 2019-01-15
+
+### Added
+
 - Add "CHANGELOG.md" file  
   
 ### Changed
+
+### Fixed
 
 ### Removed